refactor: Move contract handlers to the wallet context
